Use VSCode as the editor
So now the .gitconfig file is set up to use VSCode as the editor.

Also, of note, this is the last day of the Copilot trial; I think it ends, however I have attempted to keep it working by setting up a shell business account on GH. Who knows. Anyway, where was I?

[Ticket: X]